[PL] Class Project moved to sperate file
